dashboard/requirements.txt
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0

